requirements.txt
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
anyio==4.10.0
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
Brotli==1.1.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
clerk-sdk-python==0.1.0
clerk_django==1.0.3
cryptography==45.0.6
Deprecated==1.2.18
dj-database-url==3.0.1
Django==5.2.5
django-clerk==0.1.15
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
ecdsa==0.19.1
exceptiongroup==1.3.0
frozenlist==1.7.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
multidict==6.6.4
packaging==25.0
pillow==11.3.0
propcache==0.3.2
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycares==4.10.0
pycparser==2.22
pydantic==1.10.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
svix==1.56.0
types-Deprecated==1.2.15.20250304
types-python-dateutil==2.9.0.20250822
typing_extensions==4.14.1
urllib3==2.5.0
whitenoise==6.9.0
wrapt==1.17.3
yarl==1.20.1
